+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "that-api-events",
-  "version": "2.8.0",
+  "version": "2.9.0",
   "description": "THATConference.com events service",
   "main": "index.js",
   "engines": {
@@ -23,6 +23,7 @@
     "@apollo/federation": "0.22.0",
     "@fast-csv/format": "^4.3.5",
     "@google-cloud/firestore": "^5.0.1",
+    "@google-cloud/trace-agent": "^5.1.6",
     "@graphql-tools/load-files": "6.2.7",
     "@graphql-tools/merge": "6.2.10",
     "@sentry/node": "^6.16.0",
